add message REST API
The message endpoint only exists under /projects, since it is not
listable--message URIs can be discovered via the series REST API, or via
the /api/v1/projects/NN/messages/ID/replies route.  Compared to the
information under /series, it lets you retrieve the full mbox contents
(and, when plugins support will be added, the tags that are parsed from
replies).  The mbox can also be retrieved as a simple text/plain response.