mods: refactor extraction of configuration into a dictionary
Both the email and testing modules have code to extract the
configuration from project properties into a dictionary.
Generalize that code, using the project_config_schema to
drive the conversion, and use it in the git plugin as well

This matches the way configuration will be stored in the database
when we move away from project properties.  In fact, all
this nice visitor code will disappear very soon...

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>