models: store plugin configuration in a single, separate JSONField
This is the next step in the splitting the properties table into
logically separate components.  This time it's the turn of configuration
data coming from the plugin schemas, which is extracted into a new
JSON field.

The configuration editor is simplified a bit because all the properties
are sent to the server in a single shot and all configuration is
overwritten.  As a result, we no longer need the deletion API
delete-project-properties-by-prefix.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>