models: switch from property tables to field
Fortunately, most accesses to the property tables were hidden behind
accessors.  Therefore, apart from the usual ugly migration (which in
this case is more or less a copy of the config migration) it is
almost enough to change the accessors to look into the JSON dictionary.

We can also use the new API to access a whole subset of the properties,
which simplifies the access to testing.check_in.*.